PAY-762 - fixed test failure and just file rename
